requirements.txt
torch==2.3.1
wandb==0.17.4
numpy==2.0.0
pandas==2.2.2
transformers==4.42.4
pillow==10.4.0
datasets==2.20.0
torchvision==0.18.1
python-dotenv==1.0.1
matplotlib==3.9.1
tqdm==4.66.4
